:white_check_mark: fix: fix e2e test
