+++ e2e/__tests__/basic/image.e2e.ts
@@ -1,6 +1,10 @@
-import { initHtml2Sketch, isUpdate, outputJSONData } from './utils';
-import inlineImage from './json/inline-image.json';
-import pngUrl from './json/png-url-image.json';
+import {
+  initHtml2Sketch,
+  isUpdate,
+  outputJSONData,
+  pngURLImageJSON,
+  inlineImageJSON,
+} from '@e2e';
 
 describe('测试图片', () => {
   test('inline 的图片类型正常', async () => {
@@ -18,8 +22,8 @@ describe('测试图片', () => {
     if (isUpdate) {
       outputJSONData(image, 'inline-image');
     }
-    expect(image).toStrictEqual(inlineImage);
-  });
+    expect(image).toStrictEqual(inlineImageJSON);
+  }, 30000);
 
   test('PNG图片链接类型正常', async () => {
     const html2Sketch = await initHtml2Sketch();
@@ -36,6 +40,6 @@ describe('测试图片', () => {
     if (isUpdate) {
       outputJSONData(image, 'png-url-image');
     }
-    expect(image).toStrictEqual(pngUrl);
+    expect(image).toStrictEqual(pngURLImageJSON);
   }, 30000);
 });
